+++ src/parser/inline/comment.test.ts
@@ -16,6 +16,7 @@ describe('Unit: parser/inline/comment', () => {
       assert.deepStrictEqual(inspect(parser('<#\\ a #>')), undefined);
       assert.deepStrictEqual(inspect(parser('<# a ##>')), undefined);
       assert.deepStrictEqual(inspect(parser('<## a #>')), undefined);
+      assert.deepStrictEqual(inspect(parser('<!- ->')), undefined);
     });
 
     it('ab', () => {
@@ -28,6 +29,7 @@ describe('Unit: parser/inline/comment', () => {
       assert.deepStrictEqual(inspect(parser('<## a ##>')), [[], '']);
       assert.deepStrictEqual(inspect(parser('<# #>a')), [[], 'a']);
       assert.deepStrictEqual(inspect(parser('<# a #>b')), [[], 'b']);
+      assert.deepStrictEqual(inspect(parser('<!-- -->')), [[], '']);
     });
 
   });